+++ src/Core/IncomeCalculator/CalculateStrategy/FinalAmountWithoutRegularPayments.php
@@ -27,7 +27,7 @@ class FinalAmountWithoutRegularPayments implements CalculateStrategyInterface
             (float)$input->getInitialAmount(),
             (float)$input->getRegularPayment(),
             $input->getNumberOfRegularPaymentsPerYear(),
-            (float)$input->getNumberOfYears(),
+            (int)$input->getNumberOfYears(),
             (float)$input->getInterestRatePerYear(),
             $finalAmount
         );
