+++ src/Core/SpendCalculator/Model/Input.php
@@ -12,13 +12,13 @@ class Input
 
     private int $numberOfPaymentsPerYear;
 
-    private ?float $numberOfYears;
+    private ?int $numberOfYears;
 
     private ?float $interestRatePerYear;
 
     private ?float $finalAmount;
 
-    private float $numberOfYearsUntilFistPayment;
+    private int $numberOfYearsUntilFistPayment;
 
     private float $inflation;
 
@@ -32,10 +32,10 @@ class Input
         ?float $initialAmount,
         ?float $paymentAmount,
         int $numberOfPaymentsPerYear,
-        ?float $numberOfYears,
+        ?int $numberOfYears,
         ?float $interestRatePerYear,
         ?float $finalAmount,
-        float $numberOfYearsUntilFistPayment,
+        int $numberOfYearsUntilFistPayment,
         float $inflation,
         bool $initialAmountIsUnknown,
         bool $paymentAmountIsUnknown,
@@ -81,7 +81,7 @@ class Input
         return $this->numberOfPaymentsPerYear;
     }
 
-    public function getNumberOfYears(): ?float
+    public function getNumberOfYears(): ?int
     {
         if ($this->numberOfYearsIsUnknown()) {
             return null;
@@ -108,7 +108,7 @@ class Input
         return $this->finalAmount;
     }
 
-    public function getNumberOfYearsUntilFistPayment(): float
+    public function getNumberOfYearsUntilFistPayment(): int
     {
         return $this->numberOfYearsUntilFistPayment;
     }
